requirements.txt
fastapi
uvicorn
pandas
transformers
torch
sentence-transformers
rank-bm25
scikit-learn
faiss-cpu
peft
requests
langchain
langchain-community
langchain-core
langchain-huggingface
sec-edgar-downloader
PyYAML>=6.0
langchain>=0.2.10
langchain-core>=0.2.30
langchain-community>=0.2.10
pyyaml>=6.0.1
faiss-cpu>=1.8.0
